+++ R/colcheck.R
@@ -18,7 +18,7 @@
 #' biomass_rec <- recipe(HHV ~ ., data = biomass) %>%
 #'    step_rm(sample, dataset) %>%
 #'    check_cols(contains("gen")) %>%
-#'    step_center(all_predictors())
+#'    step_center(all_numeric_predictors())
 #'
 #' \dontrun{
 #' bake(biomass_rec, biomass[, c("carbon", "HHV")])